enpoint to create new productos is added

+++ server/app/types/company.go
@@ -1,6 +1,6 @@
 package types
 
 type Company struct {
-	Id   *int   `json:"id"`
+	Id   *int64 `json:"id"`
 	Name string `json:"name"`
 }
